+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 決済サービスのserviceOutputをArrayに対応
+- PermitのissuedThrough.typeOfをCreditCardに対応
 
 ### Deprecated